+++ tests/Pest.php
@@ -1,0 +1,5 @@
+<?php
+
+use Syriable\Filament\Plugins\Translator\Tests\TestCase;
+
+uses(TestCase::class)->in('Feature');